dm2nsc/requirements.txt
requests
arrow
cloudscraper
python-dotenv
certifi